refactor: Player와 Dealer의 중복된 코드를 추상클래스로 제거
